Add Object#set_singleton_class
This removes the object's previous singleton class and included
modules, and makes the given class the object's previous class.
Any modules included in the given class are now modules that
extend the object.  However, the object's main class does not
change, so it does not inherit methods in the superclass of
the given class.

Add evilr__next_class for getting the next class in the
method lookup chain, skipping iclasses.  Add
evilr__iclass_before_next_class for getting the last iclass before
the next class in the method lookup chain.