Add Class#inherit, for inheriting from multiple classes
inherit basically converts the classes to modules (including
modules in those classes, but not their superclasses).

This commit fixes an important bug in Class#to_module by
not modifying the existing method chain for a class.  Previously,
a class with an included module that called to module would
end up with no superclass, because it modified the super pointer
of the last included iclass.  Now, it creates a copy of the chain,
by including all included modules in reverse order.

Add evilr__check_compatible_classes to DRY up some code.

Add evilr_include_iclasses for copying the iclass chain in
reverse order, using recursion.