Add Object#extend_between
Similar to Module#include_between in how it works with module
insertion, but similar to Object#unextend in that it does not
go beyond the object's class.

Also, fix an issue with Object#unextend for objects without
existing singleton classes.

Add evilr__iclass_matching_before, to returning an iclass
without going above a given node in the super chain.