Add Object#swap for completely swapping objects
Swapping classes with non-classes and modules with non-modules
is forbidden because it is really easy to get a segfault due
to the super chain.

This uses variable sized arrays, which are forbidden in C90,
but evil code doesn't care about standards.

Switch swap_instance_variables to use a temporary array instead
of a new object, for potentially more robust garbage collection.