Fix Object#swap_instance_variables when used with modules
Remove the ROBJECT_IV_INDEX_TBL macro that was defined for
1.8, since it wasn't the same as the one on 1.9.  Instead,
rename the 1.8 macro to ROBJECT_IVPTR, though it still
doesn't mean the same thing as in 1.9.

On 1.8, you can swap a module's instance variables with
a regular object, but that's not simple on 1.9, so we'll
just disallow it on both.

On 1.9, fix it so swap_instance_variables doesn't swap
the super, method table, and iv_index_tbl for the objects,
as that would cause problems.