Add ruby 1.8.6 support
ruby 1.8.6 didn't define some of the macros evilr used, so
do the ifndef define endif dance for those.

Also, parsing changed slightly between 1.8.6 and 1.8.7 that
led to a couple spec failures, so add parantheses to fix that.

Also, all singleton classes should have an attached object, so
raise an error instead of doing some wrong if that isn't the
case.