Add support for true multiple inheritance to Class#inherit
Before, Class#inherit would make a copy of the class as a module
and just include that copy.  That had a number of issues compared
to how normal ruby inheritance works:

1) The copy was separate from the class, so modifications to
the original class had no effect.

2) Superclasses were not copied.

3) The singleton classes of the inherited classes did not get
inserted into the super chain of the singleton class of the
receiver of inherit.

This is a much more involved approach that uses the original
class method tables, handles ancestry correctly by copying all
superclasses up to the first common ancestor, and works with
singleton classes.

While this passes all specs on 1.8.7, it currently has 4 failing
specs on 1.9.2. The failing specs all relate to the usage of super
when multiple classes are inherited by the same class.  Instead of
walking the super chain (which I verified is correct), super
appears to be jumping directly to the definition in the original
parent class (if one exists).  That's a completely sane operation
in normal ruby code, since a class is not supposed to be able
to change its superclass, but I'm not sure it's possible to work
around this issue easily in evilr.